chore: add CodeEditor component
